+++ src/event-listeners-handler/EventListenersHandler.test.ts
@@ -7,6 +7,7 @@ import UserActionHandler from '../user-action/UserActionHandler'
 import { nop } from '../utils/nop'
 import UserActionsHistory from '../user-actions-history/UserActionsHistory'
 import MemorySessionIdHandler from '../session-id-handler/MemorySessionIdHandler'
+import { MemorySessionSizeController } from '../session-size-controller/MemorySessionSizeController'
 
 describe('EventListenersHandler', () => {
   let eventListenersHandler: EventListenersHandler
@@ -18,7 +19,14 @@ describe('EventListenersHandler', () => {
   beforeEach(() => {
     userActionsHistory = new MemoryUserActionsHistory()
     const sessionIdHandler = new MemorySessionIdHandler(() => 'aaa-111', 500)
-    userActionHandler = new UserActionHandler(sessionIdHandler, 0, nop, nop, userActionsHistory)
+    userActionHandler = new UserActionHandler(
+      sessionIdHandler,
+      0,
+      new MemorySessionSizeController(1),
+      nop,
+      nop,
+      userActionsHistory,
+    )
 
     clickEventListener = new ClickEventListener(userActionHandler, window)
     beforeUnloadEventListener = new BeforeUnloadEventListener(userActionHandler, window)