+++ src/event-listeners/KeyDownEventListener.test.ts
@@ -9,6 +9,7 @@ import MemorySessionIdHandler from '../session-id-handler/MemorySessionIdHandler
 import KeyDownEventListener from './KeyDownEventListener'
 import * as createTargetedUserActionModule from '../user-action/createTargetedUserAction'
 import ISessionIdHandler from '../session-id-handler/ISessionIdHandler'
+import { MemorySessionSizeController } from '../session-size-controller/MemorySessionSizeController'
 
 describe('KeyDownEventListener', () => {
   let userActionHistory: UserActionsHistory
@@ -21,7 +22,14 @@ describe('KeyDownEventListener', () => {
     vitest.restoreAllMocks()
     userActionHistory = new MemoryUserActionsHistory()
     sessionIdHandler = new MemorySessionIdHandler(() => 'aaa-111', 500)
-    userActionHandler = new UserActionHandler(sessionIdHandler, 0, nop, nop, userActionHistory)
+    userActionHandler = new UserActionHandler(
+      sessionIdHandler,
+      0,
+      new MemorySessionSizeController(1),
+      nop,
+      nop,
+      userActionHistory,
+    )
     handleSpy = vitest.spyOn(userActionHandler, 'handle')
     createTargetedUserActionSpy = vitest.spyOn(createTargetedUserActionModule, 'createTargetedUserAction')
   })