+++ src/event-listeners/KeyUpEventListener.test.ts
@@ -6,6 +6,7 @@ import createElementInJSDOM from '../test-utils/createElementInJSDOM'
 import KeyUpEventListener from '../event-listeners/KeyUpEventListener'
 import MemorySessionIdHandler from '../session-id-handler/MemorySessionIdHandler'
 import * as createTargetedUserActionModule from '../user-action/createTargetedUserAction'
+import { MemorySessionSizeController } from '../session-size-controller/MemorySessionSizeController'
 
 describe('KeyUpEventListener', () => {
   let sessionIdHandler: MemorySessionIdHandler
@@ -17,7 +18,7 @@ describe('KeyUpEventListener', () => {
     beforeEach(() => {
       vitest.restoreAllMocks()
       sessionIdHandler = new MemorySessionIdHandler(() => 'aaa-111', 500)
-      userActionHandler = new UserActionHandler(sessionIdHandler, 0, nop)
+      userActionHandler = new UserActionHandler(sessionIdHandler, 0, new MemorySessionSizeController(1), nop)
       handleSpy = vitest.spyOn(userActionHandler, 'handle')
       createTargetedUserActionSpy = vitest.spyOn(createTargetedUserActionModule, 'createTargetedUserAction')
     })
@@ -26,7 +27,7 @@ describe('KeyUpEventListener', () => {
       const { element, domWindow } = createElementInJSDOM(
         `
             <div>
-                <input id="text-5" type="search" />
+                <input id='text-5' type='search' />
             </div>`,
         'div',
       )
@@ -45,7 +46,7 @@ describe('KeyUpEventListener', () => {
       const { element, domWindow } = createElementInJSDOM(
         `
             <div>
-                <input id="text-5" type="search" />
+                <input id='text-5' type='search' />
             </div>`,
         'div',
       )
@@ -69,7 +70,7 @@ describe('KeyUpEventListener', () => {
       const { element, domWindow } = createElementInJSDOM(
         `
                 <div>
-                    <input type="checkbox" id="checkbox1" name="checkbox1"/>
+                    <input type='checkbox' id='checkbox1' name='checkbox1'/>
                 </div>`,
         'div',
       )
@@ -88,7 +89,7 @@ describe('KeyUpEventListener', () => {
       const { element, domWindow } = createElementInJSDOM(
         `
                 <div>
-                    <div role="cell"/>
+                    <div role='cell'/>
                 </div>`,
         'div',
       )
@@ -107,9 +108,9 @@ describe('KeyUpEventListener', () => {
       const { element, domWindow } = createElementInJSDOM(
         `
             <div>
-                <input id="text-1" type="text"/>
-                <textarea id="text-2"></textarea>
-                <input id="text-5" type="search" />
+                <input id='text-1' type='text'/>
+                <textarea id='text-2'></textarea>
+                <input id='text-5' type='search' />
             </div>`,
         'div',
       )
@@ -148,9 +149,9 @@ describe('KeyUpEventListener', () => {
         const { element, domWindow } = createElementInJSDOM(
           `
             <div>
-                <input id="text-1" type="text"/>
-                <textarea id="text-2"></textarea>
-                <input id="text-5" type="search" />
+                <input id='text-1' type='text'/>
+                <textarea id='text-2'></textarea>
+                <input id='text-5' type='search' />
             </div>`,
           'div',
         )