+++ src/types.ts
@@ -99,6 +99,7 @@ export interface CollectorOptions {
   customSelector?: string
   sessionsPercentageKept: number
   rejectSession: () => boolean
+  minimumUserActions: number
   onPublish?: (userActions: SessionUserAction[]) => void
 }