+++ src/utils/completeOptions.test.ts
@@ -36,6 +36,7 @@ describe('completeOptions', () => {
         excludeRegex: null,
         sessionsPercentageKept: 100,
         rejectSession: DEFAULT_SESSION_REJECTION,
+        minimumUserActions: 1,
       }
       expect(completed).toStrictEqual(expected)
     })
@@ -47,6 +48,7 @@ describe('completeOptions', () => {
         maxDelay: 15,
         sessionsPercentageKept: 33.3,
         rejectSession: customSessionRejection,
+        minimumUserActions: 42,
       })
       const expected: CollectorOptions = {
         authKey: '',
@@ -57,6 +59,7 @@ describe('completeOptions', () => {
         excludeRegex: null,
         sessionsPercentageKept: 33.3,
         rejectSession: customSessionRejection,
+        minimumUserActions: 42,
       }
       expect(completed).toStrictEqual(expected)
     })
@@ -82,6 +85,7 @@ describe('completeOptions', () => {
           excludeRegex: null,
           sessionsPercentageKept: 100,
           rejectSession: DEFAULT_SESSION_REJECTION,
+          minimumUserActions: 1,
         }
         expect(completed).toStrictEqual(expected)
       })
@@ -101,6 +105,7 @@ describe('completeOptions', () => {
           excludeRegex: null,
           sessionsPercentageKept: 100,
           rejectSession: DEFAULT_SESSION_REJECTION,
+          minimumUserActions: 1,
         }
         expect(completed).toStrictEqual(expected)
       })
@@ -120,6 +125,7 @@ describe('completeOptions', () => {
           excludeRegex: null,
           sessionsPercentageKept: 100,
           rejectSession: DEFAULT_SESSION_REJECTION,
+          minimumUserActions: 1,
         }
         expect(completed).toStrictEqual(expected)
       })
@@ -139,6 +145,7 @@ describe('completeOptions', () => {
           excludeRegex: /^#my-id-.*$/,
           sessionsPercentageKept: 100,
           rejectSession: DEFAULT_SESSION_REJECTION,
+          minimumUserActions: 1,
         }
         expect(completed).toStrictEqual(expected)
       })
@@ -158,6 +165,7 @@ describe('completeOptions', () => {
           excludeRegex: null,
           sessionsPercentageKept: 100,
           rejectSession: DEFAULT_SESSION_REJECTION,
+          minimumUserActions: 1,
         }
         expect(completed).toStrictEqual(expected)
       })